+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
+* Added `AlertMethods`: Alemba vFire, Tippingpoint SMS [275](https://github.com/greenbone/python-gvm/pull/275)
+* Added `AlertConditions`: Error, SeverityChanged [275](https://github.com/greenbone/python-gvm/pull/275)
+* Added `AlertEvents`: Assigned ticket changed, Owned ticket changed, Ticket received [275](https://github.com/greenbone/python-gvm/pull/275)
+
 ### Changed
 ### Deprecated
 ### Removed